refactor: safe background modifier
